Extracts downgrade logic to script, fixes ts-node conflicts with prisma
