--- apps/web/@prisma/client.ts
@@ -1,1 +1,0 @@
-export * from "@calcom/prisma/client";
