+++ apps/web/ee/lib/stripe/customer.ts
@@ -1,9 +1,10 @@
 import { Prisma } from "@prisma/client";
 
-import stripe from "@ee/lib/stripe/server";
+import prisma from "@calcom/prisma";
 
 import { HttpError as HttpCode } from "@lib/core/http/error";
-import { prisma } from "@lib/prisma";
+
+import stripe from "./server";
 
 export async function getStripeCustomerFromUser(userId: number) {
   // Get user
