+++ packages/tsconfig/base.json
@@ -18,6 +18,9 @@
   },
   "exclude": ["node_modules"],
   "ts-node": {
+    "files": true,
+    "require": ["tsconfig-paths/register"],
+    "experimentalResolverFeatures": true,
     "compilerOptions": {
       "module": "CommonJS",
       "types": ["node"]
