+++ packages/tsconfig/package.json
@@ -7,5 +7,8 @@
     "base.json",
     "nextjs.json",
     "react-library.json"
-  ]
+  ],
+  "devDependencies": {
+    "tsconfig-paths": "^3.12.0"
+  }
 }
